# swim_coach_plugin_plan/contracts/openapi-skeleton.yaml
openapi: 3.1.0
info:
  title: Swim Coach API
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - cookieSession: []
paths:
...
  /me:
    get:
      operationId: getMe
      responses:
        "200": {description: Current user profile}
        "401": {$ref: "#/components/responses/Problem"}
  /me/profile:
    patch:
      operationId: updateProfile
      parameters: [{$ref: "#/components/parameters/CsrfToken"}]
      responses:
        "200": {description: Profile updated with optimistic version control}
        "409": {$ref: "#/components/responses/Problem"}
  /pools:
    get:
      operationId: listPools
      responses: {"200": {description: User-owned pools}}
    post:
      operationId: createPool
      parameters:
        - {$ref: "#/components/parameters/CsrfToken"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      responses:
        "201": {description: User-owned pool created}
        "409": {$ref: "#/components/responses/Problem"}
  /pools/{pool_id}:
    patch:
      operationId: updatePool
      parameters:
        - {in: path, name: pool_id, required: true, schema: {type: string, format: uuid}}
        - {$ref: "#/components/parameters/CsrfToken"}
      responses:
        "200": {description: User-owned pool updated}
        "404": {$ref: "#/components/responses/Problem"}
  /availability:
    get:
      operationId: listAvailability
      responses: {"200": {description: User availability rules}}
    put:
      operationId: replaceAvailability
      parameters: [{$ref: "#/components/parameters/CsrfToken"}]
      responses: {"200": {description: Availability replaced atomically}}
  /goals:
    get:
      operationId: listGoals
      responses: {"200": {description: User-owned goals}}
    post:
      operationId: createGoal
      parameters:
        - {$ref: "#/components/parameters/CsrfToken"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      responses: {"201": {description: Goal created with derived pace}}
  /goals/{goal_id}:
    patch:
      operationId: updateGoal
      parameters:
        - {in: path, name: goal_id, required: true, schema: {type: string, format: uuid}}
        - {$ref: "#/components/parameters/CsrfToken"}
      responses:
        "200": {description: Goal updated and pace recalculated}
        "404": {$ref: "#/components/responses/Problem"}
...
components:
  securitySchemes:
    cookieSession: {type: apiKey, in: cookie, name: swim_coach_session}
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    CsrfToken:
      in: header
      name: X-CSRF-Token
      required: true
      schema: {type: string, minLength: 32, maxLength: 200}
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema: {type: string, minLength: 8, maxLength: 200}
  responses:
    Problem:
      description: Problem Details
      content:
        application/problem+json:
          schema:
            type: object
            required: [type, title, status, code, correlation_id]
            properties:
              type: {type: string}
              title: {type: string}
              status: {type: integer}
              detail: {type: string}
              code: {type: string}
              correlation_id: {type: string}